refactor: extract APP-23 graph time-window helpers into src/card-graph-view.js
Pure mechanical relocation of app.js's APP-23 chapter (GV_WIN_OPTS, loadGvWin,
saveGvWin, gvWinLabel, gvWinCutoff, gvBuckets) into its own ES module. This
chapter is stamped RETIRED (its per-card tile dashboard was replaced by the
Round-Up board), but the time-window/bucket helpers survive and are still
called by the Graph Carousel (APP-24, still in app.js). No behavior change.

Left a pointer banner at the old location so downstream APP-NN ids keep their
position. Added the new file to the code-map generator and the service
worker's offline SHELL precache.

Gates green: smoke, logic-test (706/706), lease/lease-deploy/promote/cachebust
suites, gen-rule-usage --check, check-window-catalog, gen-code-map --check.
